% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smith_fretwell_functions.R
\name{add_to_sf_plot}
\alias{add_to_sf_plot}
\title{Add a new Smith-Fretwell curve to a ggplot object created with make_sf_plot}
\usage{
add_to_sf_plot(
  existing_plot,
  Iymin = 6,
  Womax = 0.6,
  Ip = 1000,
  shapeval = 0.02
)
}
\arguments{
\item{existing_plot}{ggplot object created with make_sf_plot that has the first (base) curve}

\item{Iymin}{Minimum seed size for a viable offspring}

\item{Womax}{Maximum fitness that an offspring can achieve}

\item{Ip}{Total amount of energy an adult has to invest into reproduction}

\item{shapeval}{slope of the Smith-Fretwell curve}
}
\description{
Add a new Smith-Fretwell curve to a ggplot object created with make_sf_plot
}
