% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_sink_dynamics.R
\name{assumption_check}
\alias{assumption_check}
\title{check if input parameters violates model(Pulliam 1988) assumptions}
\usage{
assumption_check(params)
}
\arguments{
\item{params}{a vector of model parameters}
}
\description{
check if input parameters violates model(Pulliam 1988) assumptions
}
\examples{
# VALID parameter example 1: lambdaSource > 1, lambdaSink < 1
assumption_check(params = c(pa = 0.6, pj = 0.15, betaSource = 4, betaSink = 1))
# INVALID parameter example 1: lambdaSource < 1, lambdaSInk < 1
assumption_check(params = c(pa = 0.6, pj = 0.15, betaSource = 2, betaSink = 2))
# INVALID parameter example 2: lambdaSource < 1, lambdaSink > 1
assumption_check(params = c(pa = 0.6, pj = 0.15, betaSource = 2, betaSink = 3))
}
\keyword{internal}
