% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotic_resource_competition.R
\name{biotic_competition}
\alias{biotic_competition}
\title{Internal function for running the biotic resource
competition between two consumers and a shared resource species}
\usage{
biotic_competition(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of the two consumers P1 and P2,
and the resource species H}

\item{params}{vector of parameters:
r = per capita growth rate of the resource species
q = 1/prey carrying capacity of the resource species H
T_h1, T_h2 = handling time for consumer species P1 and P2
a1, a2 = attack rate of consumer species P1 and P2
e1, e2 = conversion efficiency of consumer species P1 and P2
d1, d2 = predator death rate of consumer species P1 and P2}
}
\description{
Internal function for running the biotic resource
competition between two consumers and a shared resource species
}
\keyword{internal}
