% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{discretelogistic_eqn}
\alias{discretelogistic_eqn}
\title{Equation that returns Nt+1 from Nt for the "standard" Discrete Logistic model}
\usage{
discretelogistic_eqn(Nt, rd, K)
}
\arguments{
\item{Nt}{Nt}

\item{rd}{Population growth rate}

\item{K}{Carrying capacity}
}
\description{
Equation that returns Nt+1 from Nt for the "standard" Discrete Logistic model
}
\keyword{internal}
