% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoevoapps.R
\docType{package}
\name{ecoevoapps-package}
\alias{ecoevoapps}
\alias{ecoevoapps-package}
\title{ecoevoapps: Simulate Dynamics of Ecology/Evolution Models}
\description{
A collection of Shiny apps for common models in ecology and evolution. The package also includes functions that can be called directly from R (rather than from the apps).
}
\author{
\strong{Maintainer}: Gaurav Kandlikar \email{gkandlikar@missouri.edu}

Authors:
\itemize{
  \item Madeline Cowen \email{mcowen@ucla.edu}
  \item Kenji Hayashi \email{kthayashi@ucla.edu}
  \item Rosa McGuire \email{rmcguire1@ucla.edu}
  \item Xinyi Yan \email{xinyiyan@utexas.edu}
  \item Marcel Vaz \email{marcelvaz@gmail.com}
}

Other contributors:
\itemize{
  \item Alejandra Martinez-Blancas \email{alemtzb@ciencias.unam.mx} [contributor]
  \item Damla Cinoglu \email{dcinoglu@utexas.edu} [contributor]
  \item Sheila Vazquez-Morales \email{sheila.vazquez-morales@health.missouri.edu} [contributor]
}

}
\keyword{internal}
