% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_continuous.R
\name{exponential_growth}
\alias{exponential_growth}
\title{Exponential growth model}
\usage{
exponential_growth(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{intrinsic growth rate r}
}
\description{
Exponential growth model
}
\keyword{internal}
