% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibiogeo_functions.R
\name{ibiogeo_E}
\alias{ibiogeo_E}
\title{Extinction rate for Island Biogeography model}
\usage{
ibiogeo_E(S, A, k)
}
\arguments{
\item{S}{number of species on island}

\item{A}{size of island (km^2)}

\item{k}{scaling constant}
}
\description{
Extinction rate for Island Biogeography model
}
