% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibiogeo_functions.R
\name{ibiogeo_Ex}
\alias{ibiogeo_Ex}
\title{equilibrium extinction rate for Island Biogeography model}
\usage{
ibiogeo_Ex(D, A, M, k)
}
\arguments{
\item{D}{distance from mainland (km)}

\item{A}{size of island (km^2)}
}
\description{
equilibrium extinction rate for Island Biogeography model
}
\keyword{internal}
