% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibiogeo_functions.R
\name{ibiogeo_I}
\alias{ibiogeo_I}
\title{Immigration rate for Island Biogeography model}
\usage{
ibiogeo_I(S, D, M, k)
}
\arguments{
\item{S}{number of species on island}

\item{D}{distance from mainland (km)}

\item{M}{number of species in the mainland}

\item{k}{scaling constant}
}
\description{
Immigration rate for Island Biogeography model
}
