% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibiogeo_functions.R
\name{ibiogeo_Sx}
\alias{ibiogeo_Sx}
\title{Equilibrium number of species on island for Island Biogeography model}
\usage{
ibiogeo_Sx(D, A, M, k)
}
\arguments{
\item{D}{distance from mainland (km)}

\item{A}{size of island (km^2)}

\item{M}{number of species in the mainland}

\item{k}{scaling constant}
}
\description{
Equilibrium number of species on island for Island Biogeography model
}
