% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_continuous.R
\name{lagged_logistic_growth}
\alias{lagged_logistic_growth}
\title{Lagged logistic growth model}
\usage{
lagged_logistic_growth(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{vector of r (intrinsic growth rate), K (carrying capacity), and tau (time lag)}
}
\description{
Lagged logistic growth model
}
