% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_continuous.R
\name{logistic_growth}
\alias{logistic_growth}
\title{Logistic growth model}
\usage{
logistic_growth(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{vector of r (intrinsic growth rate) and K (carrying capacity)}
}
\description{
Logistic growth model
}
