% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka_volterra_competition.R
\name{lotka_volterra_competition}
\alias{lotka_volterra_competition}
\title{Lotka-Volterra Competition with carrying capacities}
\usage{
lotka_volterra_competition(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{vector of initial population sizes for both species}

\item{params}{vector of model parameters}
}
\description{
Lotka-Volterra Competition with carrying capacities
}
\keyword{internal}
