% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_resource_variants.R
\name{lv_predprey_logPrey}
\alias{lv_predprey_logPrey}
\title{Run L-V Pred prey model, Type I FR + logistic prey}
\usage{
lv_predprey_logPrey(time, init, pars)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{pars}{intrinsic growth rate r, a, e, d, K}
}
\description{
Run L-V Pred prey model, Type I FR + logistic prey
}
