% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_resource_variants.R
\name{lv_predprey_t1}
\alias{lv_predprey_t1}
\title{Run L-V Pred prey model, Type I Functional Response}
\usage{
lv_predprey_t1(time, init, pars)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{pars}{intrinsic growth rate r, a, e, d}
}
\description{
Run L-V Pred prey model, Type I Functional Response
}
