% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predator_prey_dynamics.R
\name{lv_predprey_t2}
\alias{lv_predprey_t2}
\title{Run L-V Pred prey model, Type II Functional Response}
\usage{
lv_predprey_t2(time, init, pars)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{pars}{intrinsic growth rate r, a, e, d, T_h}
}
\description{
Run L-V Pred prey model, Type II Functional Response
}
\keyword{internal}
