% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibiogeo_functions.R
\name{make_equilibrium_plot}
\alias{make_equilibrium_plot}
\title{make equilibrium plot for ibiogeo model}
\usage{
make_equilibrium_plot(D, A, M, k)
}
\arguments{
\item{D}{distance from mainland (km)}

\item{A}{size of island (km^2)}

\item{M}{Mainland species richness}

\item{k}{scaling constant}
}
\description{
make equilibrium plot for ibiogeo model
}
