% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibiogeo_functions.R
\name{make_islands_map}
\alias{make_islands_map}
\title{make map for ibiogeo model}
\usage{
make_islands_map(tx, ty, txa, tya, txb, tyb, D, A, k, M)
}
\arguments{
\item{D}{distance from mainland (km)}

\item{A}{size of island (km^2)}
}
\description{
make map for ibiogeo model
}
