% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smith_fretwell_functions.R
\name{make_sf_plot}
\alias{make_sf_plot}
\title{Make a Smith-Fretwell plot to show the optimal seed size}
\usage{
make_sf_plot(Iymin = 3, Womax = 0.6, Ip = 1000, shapeval = 0.02)
}
\arguments{
\item{Iymin}{Minimum seed size for a viable offspring}

\item{Womax}{Maximum fitness that an offspring can achieve}

\item{Ip}{Total amount of energy an adult has to invest into reproduction}

\item{shapeval}{slope of the Smith-Fretwell curve}
}
\description{
Make a Smith-Fretwell plot to show the optimal seed size
}
