% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualism.R
\name{mutualism}
\alias{mutualism}
\title{Internal function for mutualism with saturating functional response}
\usage{
mutualism(time, init, params)
}
\arguments{
\item{time}{Vector of time units over which to run model.}

\item{init}{Named vector of initial population sizes (N1, N2).}

\item{params}{Named vector of model parameters.
\itemize{
\item r1: Per capita growth rate of species 1
\item r2: Per capita growth rate of species 1
\item a12: Effect of species 2 on species 1
\item a21: Effect of species 1 on species 2
\item b1: Half-saturation constant for species 1
\item b2: Half-saturation constant for species 2
\item d1: Death rate of self-limitation for species 1
\item d2: Death rate of self-limitation for species 2
}}
}
\description{
Used internally by \code{run_mutualism()}.
}
\keyword{internal}
