% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotic_resource_competition.R
\name{plot_biotic_comp_time}
\alias{plot_biotic_comp_time}
\title{Plot population size over time for the boitic resource competition model}
\usage{
plot_biotic_comp_time(sim_df)
}
\arguments{
\item{sim_df}{matrix or data frame of simulations generated by
\link{run_biotic_comp_model}}
}
\description{
Plot population size over time for the boitic resource competition model
}
\examples{
sim_df_biotic_rc <- run_biotic_competition_model(
time = seq(0,10),
init = c(P1 = 25, P2 = 25, H = 30),
params = c(r = 0.2, q = .0066,
a1 = .02, T_h1 = 0.1, e1 = 0.4, d1 = 0.1,
a2 = .02, T_h2 = 0.1, e2 = 0.39, d2 = 0.1))
plot_biotic_comp_time(sim_df_biotic_rc)
}
\seealso{
\code{\link[=run_biotic_comp_model]{run_biotic_comp_model()}}  for simulating the biotic resource
competition model
}
