% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_continuous.R
\name{plot_continuous_population_growth}
\alias{plot_continuous_population_growth}
\title{Plot the dynamics of a discrete population growth model}
\usage{
plot_continuous_population_growth(sim_df)
}
\arguments{
\item{sim_df}{data frame of continuous population growth}
}
\description{
Plot the dynamics of a discrete population growth model
}
\examples{
# Exponential population growth
exp_params <- c(r = 0.5)
sim_df_ex <- run_exponential_model(time = 10, init = c(N1 = 10),
params = exp_params)
plot_continuous_population_growth(sim_df_ex)
log_params <- c(r = 0.5, K = 100)
sim_df_log <- run_logistic_model(time = 10, init = c(N1 = 10),
params = log_params)
plot_continuous_population_growth(sim_df_log)
}
\seealso{
\code{\link[=run_logistic_model]{run_logistic_model()}} and \code{\link[=run_exponential_model]{run_exponential_model()}} for functions
that generate the \code{sim_df} objects plotted by this function
}
