% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{plot_discrete_population_cobweb}
\alias{plot_discrete_population_cobweb}
\title{Plot a cobweb digram of a discrete population growth model}
\usage{
plot_discrete_population_cobweb(sim_df, params_vec, model_type)
}
\arguments{
\item{sim_df}{data frame of discrete population growth
generated by run_run_beverton_holt_model(), run_ricker_model(),
run_discrete_logistic_model(), run_discrete_exponential_model()}

\item{params_vec}{a vector with entries c(rd = XX and K = YY)
describing the parameters used to generate sim_df}

\item{model_type}{a string: either "discrete_logistic", "beverton_holt", or "ricker"}
}
\description{
Plot a cobweb digram of a discrete population growth model
}
\examples{
dl_params <- c(rd = 1.1, K = 1000)
sim_df_dl <- run_discrete_logistic_model(N0 = 1, K = dl_params["K"], rd = dl_params["rd"], time = 100)
plot_discrete_population_cobweb(sim_df_dl, params_vec = dl_params, model_type = "discrete_logistic")
bh_params <- c(rd = 1.1, K = 1000)
sim_df_bh <- run_beverton_holt_model(N0 = 1, K = bh_params["K"], rd = bh_params["rd"], time = 100)
plot_discrete_population_cobweb(sim_df_bh, params_vec = bh_params, model_type = "beverton_holt")
ri_params <- c(rd = 1.1, K = 1000)
sim_df_ri <- run_ricker_model(N0 = 1, K = ri_params["K"], rd = ri_params["rd"], time = 100)
plot_discrete_population_cobweb(sim_df_ri, params_vec = ri_params, model_type = "ricker")
}
