% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{plot_discrete_population_growth}
\alias{plot_discrete_population_growth}
\title{Plot the dynamics of a discrete population growth model}
\usage{
plot_discrete_population_growth(sim_df)
}
\arguments{
\item{sim_df}{data frame of discrete population growth
generated by run_run_beverton_holt_model(), run_ricker_model(),
run_discrete_logistic_model(), run_discrete_exponential_model()}
}
\description{
Plot the dynamics of a discrete population growth model
}
\examples{
sim_df <- run_beverton_holt_model(N0 = 1, K = 100, Rd = 1.01, time = 100)
plot_discrete_population_growth(sim_df)
}
