% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotic_resource_competition.R
\name{plot_functional_responses}
\alias{plot_functional_responses}
\title{Plot population size over time for the boitic resource competition model}
\usage{
plot_functional_responses(params)
}
\arguments{
\item{sim_df}{matrix or data frame of simulations generated by
\link{run_biotic_comp_model}}
}
\description{
Plot population size over time for the boitic resource competition model
}
\examples{
params_vec <- c(r = 0.2, q = .0066,
a1 = .02, T_h1 = 0.1, e1 = 0.4, d1 = 0.1,
a2 = .02, T_h2 = 0.1, e2 = 0.39, d2 = 0.1)
plot_functional_responses(params_vec)
}
\seealso{
\code{\link[=run_biotic_comp_model]{run_biotic_comp_model()}}  for simulating the biotic resource
competition model
}
