% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_ODE_variants.R
\name{plot_infectiousdisease_time}
\alias{plot_infectiousdisease_time}
\title{Plot phase portrait for infectious disease model model}
\usage{
plot_infectiousdisease_time(sim_df, model_type)
}
\arguments{
\item{sim_df}{simulated data frame generated from run_infectiousdisease_model()}
}
\description{
Plot phase portrait for infectious disease model model
}
\examples{
# Run SIR model
params_vec <- c(m = .1, beta = .01, v = .2, gamma = 0)
init_vec <- c(S = 100, I = 1, R = 0)
time_vec <- seq(0, 100, 0.1)
sir_out <- run_infectiousdisease_model(time = time_vec, init = init_vec, params = params_vec, model_type = "SIR")
plot_infectiousdisease_time(sir_out, model_out = "SIR")
}
