% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_ODE_variants.R
\name{plot_infectiousdisease_trajectories}
\alias{plot_infectiousdisease_trajectories}
\title{Plot trajectories for SIR model}
\usage{
plot_infectiousdisease_trajectories(sim_df, x_axis, y_axis)
}
\arguments{
\item{sim_df}{simulated data frame generated from run_infectiousdisease_model()}

\item{x_axis}{name of the column in `sim_df` to plot on X-axis of trajectory}

\item{y_axis}{name of the column in `sim_df` to plot on Y-axis of trajectory}
}
\description{
Plot trajectories for SIR model
}
