% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_population_growth.R
\name{plot_leslie_diagram}
\alias{plot_leslie_diagram}
\title{Generate diagram of population structure}
\usage{
plot_leslie_diagram(leslie_mat)
}
\arguments{
\item{leslie_mat}{Leslie matrix from which to generate diagram}
}
\description{
Generate diagram of population structure
}
