% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualism.R
\name{plot_mutualism_portrait}
\alias{plot_mutualism_portrait}
\title{Plot phase portrait for mutualism model}
\usage{
plot_mutualism_portrait(sim_df, vec = TRUE, traj = TRUE, ...)
}
\arguments{
\item{sim_df}{Data frame of mutualism model simulation returned by
\code{\link[=run_mutualism]{run_mutualism()}}.}

\item{vec}{Logical. If \code{TRUE} (default), a vector field will be drawn.}

\item{traj}{Logical. If \code{TRUE} (default), trajectories of population sizes in
\code{sim_df} will be drawn.}

\item{...}{
  Arguments passed on to \code{\link[=mutualism_vector_field]{mutualism_vector_field}}
  \describe{
    \item{\code{vec.density}}{Density of vectors, specified as a single numeric value
larger than 0. e.g. If \code{vec.density = 20} (default), a 20 \eqn{\times} 20
point grid evenly spaced along each axis is generated, where each grid
point represents the starting point of a vector.}
    \item{\code{vec.scale}}{Value to scale magnitude of vectors by, specified as a
single numeric value larger than 0. e.g. If \code{vec.scale = 0.1} (default) the
magnitude of each vector is 0.1 times the magnitude of the vector produced
by running \code{\link[=run_mutualism]{run_mutualism()}} for one time step given each starting point
(specified via \code{vec.density}) and \code{params} (inherited from \code{sim_df}).}
  }}
}
\value{
A ggplot object with population sizes (\code{N1}, \code{N2}) plotted against one
another as a phase portrait. Zero net growth isoclines for each species are
also plotted. Optionally, a vector field and/or trajectories of population
sizes can also be plotted.
}
\description{
Generate a phase portrait for the model of mutualism with saturating
functional response.
}
\note{
The plot generated with this function is constrained to display only the
first quadrant (i.e. values of 0 and above) of the phase plane defined by
\code{N1} and \code{N2} because negative values of \code{N} do not make biological sense. As
a consequence, when \code{sim_df} is generated using negative values for \code{r1}
and/or \code{r2}, the corresponding zero net growth isoclines may not be visible
in the plot if they lie largely beyond the plot's limits.
}
\examples{
# Plot phase portrait with vector field and population trajectories
sim <- run_mutualism()
plot_mutualism_portrait(sim)

# Plot phase portrait without vector field
sim <- run_mutualism()
plot_mutualism_portrait(sim, vec = FALSE)

# Plot phase portrait without population trajectories
sim <- run_mutualism()
plot_mutualism_portrait(sim, traj = FALSE)

# Customize vector field
sim <- run_mutualism()
plot_mutualism_portrait(sim, vec.density = 10, vec.scale = 0.2)
}
\seealso{
\code{\link[=run_mutualism]{run_mutualism()}}, \code{\link[=plot_mutualism_time]{plot_mutualism_time()}}
}
