% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualism.R
\name{plot_mutualism_time}
\alias{plot_mutualism_time}
\title{Plot population trajectories over time for mutualism model}
\usage{
plot_mutualism_time(sim_df)
}
\arguments{
\item{sim_df}{Data frame of mutualism model simulation returned by
\code{\link[=run_mutualism]{run_mutualism()}}.}
}
\value{
A ggplot object with trajectories of population sizes (\code{N1}, \code{N2}) plotted
against time. Population carrying capacities for each species, calculated
from model parameters, are also plotted.
}
\description{
Generate a plot of trajectories of population sizes (\code{N1}, \code{N2}) simulated
through time for the model of mutualism with saturating functional response
(Holland 2015).
}
\details{
The plot generated with this function is constrained to display only the
first quadrant of the plane defined by \code{N} and \code{time} (i.e. values of 0 and
above) because negative values of \code{N} or \code{time} do not make biological sense.
As a consequence, when \code{sim_df} is generated using negative values for \code{r1}
and/or \code{r2}, this function may generate warning messages that look like:
\enumerate{
\item \verb{Removed n rows containing missing values (geom_hline). }
\item \verb{Removed n row(s) containing missing values (geom_path). }
}

These warnings are generated by \code{ggplot2} when (1) population carrying
capacities are calculated to be negative values or (2) negative population
size values are present in \code{sim_df}, respecitvely. Here, these warnings
typically do not mean that there is a problem with \code{sim_df}, but rather are
reporting that lines (or portions of lines) with negative values have been
omitted from the plot. Inspect \code{sim_df} to further diagnose these warnings.
}
\examples{
sim <- run_mutualism()
plot_mutualism_time(sim)
}
\seealso{
\code{\link[=run_mutualism]{run_mutualism()}}, \code{\link[=plot_mutualism_portrait]{plot_mutualism_portrait()}}
}
