% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_resource_variants.R
\name{plot_predprey_time}
\alias{plot_predprey_time}
\title{Function for plotting phase portrait of predator-prey model}
\usage{
plot_predprey_time(sim_df)
}
\arguments{
\item{sim_df}{data frame generated from run_XXX}
}
\description{
Function for plotting phase portrait of predator-prey model
}
\examples{
# Run Rosenzweig-Macarthur model:
param_vec <- c(r = .1, a = .01, e = .01, d = .001, K = 1000, T_h = .1)
sim_df <- run_predprey_model(200, init = c(H = 10, P = 5), params = param_vec)
plot_predprey_time(sim_df = sim_df)
}
