% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_resource_variants.R
\name{plot_predprey_trajectory}
\alias{plot_predprey_trajectory}
\title{Function for plotting trajectory of predator-prey model}
\usage{
plot_predprey_trajectory(sim_df, param_vec, vectors_field = F, ...)
}
\arguments{
\item{sim_df}{data frame generated from run_XXX}

\item{vectors_field}{set to TRUE to see the vector field under the trajectory}

\item{pars_for_eq_func}{parameter values used to generate `sim_df`}
}
\description{
Function for plotting trajectory of predator-prey model
}
