% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_population_growth.R
\name{plot_structured_population_size}
\alias{plot_structured_population_size}
\title{Generate a trajectory of population size in each (st)age over time}
\usage{
plot_structured_population_size(pop_growth_matrix)
}
\arguments{
\item{pop_growth_matrix}{population growth matrix, as generated by
\code{run_structured_population_simulation} (each row is an age class and each
column is a time step)}
}
\value{
ggplot object of population trajectory over time
}
\description{
Generate a trajectory of population size in each (st)age over time
}
\seealso{
\code{\link[=run_structured_population_simulation]{run_structured_population_simulation()}} to simulate the growth of a
structured population given a Leslie matrix,
\code{\link[=plot_structured_population_agedist]{plot_structured_population_agedist()}} and
\code{\link[=plot_structured_population_lambda]{plot_structured_population_lambda()}} for plotting different aspects of the
population trajectory, and \code{\link[=plot_leslie_diagram]{plot_leslie_diagram()}} for plotting an
transition diagram based on the specified Leslie matrix
}
