% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_resource_variants.R
\name{plot_vector_field}
\alias{plot_vector_field}
\title{generates a vector field}
\usage{
plot_vector_field(sim_df, pars_for_eq_func, vec_density = 20)
}
\arguments{
\item{sim_df}{data frame generated from run_XXX}

\item{pars_for_eq_func}{parameter values used to generate `sim_df`}

\item{vec_density}{density of grid to generate}
}
\value{
a ggplot2 object
}
\description{
generates a vector field
}
