% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_population_growth.R
\name{popmat_to_df}
\alias{popmat_to_df}
\title{Converts population matrix into a long-format dataframe}
\usage{
popmat_to_df(pop_growth_matrix)
}
\arguments{
\item{pop_growth_matrix}{matrix where each row is an age class and each column
is a time step}
}
\description{
For internal use only
}
