% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{ricker_eqn}
\alias{ricker_eqn}
\title{Equation that returns Nt+1 from Nt for the Ricker model}
\usage{
ricker_eqn(Nt, r, K)
}
\arguments{
\item{Nt}{Nt}

\item{r}{Population growth rate}

\item{K}{Carrying capacity}
}
\description{
Equation that returns Nt+1 from Nt for the Ricker model
}
