% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_resource_variants.R
\name{rm_predprey}
\alias{rm_predprey}
\title{Run Rosenzweig-MacArthur model}
\usage{
rm_predprey(time, init, pars)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{pars}{intrinsic growth rate r, a, e, d, T_h, K}
}
\description{
Run Rosenzweig-MacArthur model
}
\keyword{internal}
