% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_resource_variants.R
\name{rm_predprey_1step}
\alias{rm_predprey_1step}
\title{Internal function used for generating vector field}
\usage{
rm_predprey_1step(H, P, pars)
}
\arguments{
\item{H}{density of the prey}

\item{P}{density of the predator}

\item{pars}{vector of model parameters (r,a,e,d,T_h)}
}
\description{
Internal function used for generating vector field
}
\keyword{internal}
