% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic_resource_competition.R
\name{Rstar_essential}
\alias{Rstar_essential}
\title{Calculate RStar for the 2-consumer, 2-essential resource competition}
\usage{
Rstar_essential(params)
}
\arguments{
\item{params}{vector of parameters: k11, k12, k21, k22, r1, r2, m1, m2}
}
\description{
Calculate RStar for the 2-consumer, 2-essential resource competition
}
