% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic_resource_competition.R
\name{run_abiotic_comp_rstar}
\alias{run_abiotic_comp_rstar}
\title{Calculate RStar for the 2-consumer, 2-essential resource competition}
\usage{
run_abiotic_comp_rstar(params)
}
\arguments{
\item{params}{vector of parameters: k11, k12, k21, k22, r1, r2, m1, m2}
}
\description{
Calculate RStar for the 2-consumer, 2-essential resource competition
}
