% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{run_beverton_holt_model}
\alias{run_beverton_holt_model}
\title{Beverton Holt model}
\usage{
run_beverton_holt_model(N0, K, Rd, time)
}
\arguments{
\item{N0}{initial population size of population}

\item{K}{Carrying capacity}

\item{Rd}{population growth rate}

\item{time}{Number of time steps over which to project the model}
}
\description{
Beverton Holt model
}
