% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotic_resource_competition.R
\name{run_biotic_comp_model}
\alias{run_biotic_comp_model}
\title{Run a model of competition for a biotic resource}
\usage{
run_biotic_comp_model(
  time = seq(0, 10),
  init = c(P1 = 25, P2 = 25, H = 30),
  params = c(r = 0.2, q = 0.0066, a1 = 0.02, T_h1 = 0.1, e1 = 0.4, d1 = 0.1, a2 = 0.02,
    T_h2 = 0.1, e2 = 0.39, d2 = 0.1)
)
}
\arguments{
\item{time}{vector of time units over which to run model, starting from 0.
\code{time} can also be supplied as just the total length of the simulation
(i.e. tmax)}

\item{init}{vector of initial population sizes for both consumer species,
with names P1 and P2, and resource species H}

\item{params}{vector of model parameters:
c(r, q, a1, T_h1, e1, d1, a2, T_h2, e2, d2)}
}
\description{
Run a model of competition for a biotic resource
}
\examples{
run_biotic_competition_model(
time = seq(0,10),
init = c(P1 = 25, P2 = 25, H = 30),
params = c(r = 0.2, q = .0066,
a1 = .02, T_h1 = 0.1, e1 = 0.4, d1 = 0.1,
a2 = .02, T_h2 = 0.1, e2 = 0.39, d2 = 0.1))
}
