% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{run_discrete_exponential_model}
\alias{run_discrete_exponential_model}
\title{Discrete exponential growth model}
\usage{
run_discrete_exponential_model(N0 = 1, lambda = 1.1, time = 100)
}
\arguments{
\item{N0}{initial population size of population}

\item{lambda}{discrete population growth rate}

\item{time}{Number of time steps over which to project the model}
}
\value{
A data frame with columns time and Nt
}
\description{
Discrete exponential growth model
}
\examples{
run_discrete_exponential_model(N0 = 1, lambda = 1.1, time = 10)
}
\seealso{
\code{\link[=run_exponential_growth_model]{run_exponential_growth_model()}} for simulation exponential growth
in continuous time, and see \code{\link[=run_discrete_logistic_model]{run_discrete_logistic_model()}},
\code{\link[=run_beverton_holt_model]{run_beverton_holt_model()}}, and \code{\link[=run_ricker_model]{run_ricker_model()}} for discrete time
models with population regulation
}
