% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{run_discrete_logistic_model}
\alias{run_discrete_logistic_model}
\title{Discrete logistic model}
\usage{
run_discrete_logistic_model(N0 = 1, K = 100, rd = 1.1, time = 100)
}
\arguments{
\item{N0}{initial population size of population}

\item{K}{Carrying capacity}

\item{rd}{Discrete growth factor}

\item{time}{Number of time steps over which to project the model}
}
\description{
Discrete logistic model
}
