% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_continuous.R
\name{run_exponential_model}
\alias{run_exponential_model}
\title{Run exponential growth model}
\usage{
run_exponential_model(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{intrinsic growth rate r}
}
\description{
Run exponential growth model
}
