% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibiogeo_functions.R
\name{run_ibiogeo_model}
\alias{run_ibiogeo_model}
\title{equilibrium extinction rate for Island Biogeography model}
\usage{
run_ibiogeo_model(D = c(1, 4), A = c(1, 0.5), k = 0.015, M = 100)
}
\arguments{
\item{D}{distance from mainland (km)}

\item{A}{size of island (km^2)}
}
\description{
equilibrium extinction rate for Island Biogeography model
}
