% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka_volterra_competition.R
\name{run_lvcomp_model}
\alias{run_lvcomp_model}
\title{Run the Lotka-Volterra competition model (with carrying capacity and relative effects)}
\usage{
run_lvcomp_model(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{vector of initial population sizes for both species}

\item{params}{vector of model parameters}
}
\description{
Run the Lotka-Volterra competition model (with carrying capacity and relative effects)
}
