% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualism.R
\name{run_mutualism}
\alias{run_mutualism}
\title{Run model of mutualism with saturating functional response}
\usage{
run_mutualism(
  time = 0:50,
  init = c(N1 = 20, N2 = 40),
  params = c(r1 = 0.5, r2 = 0.5, a12 = 0.8, a21 = 0.4, b1 = 10, b2 = 10, d1 = 0.02, d2
    = 0.01)
)
}
\arguments{
\item{time}{Vector of time units over which to run the model, starting at 0.
\code{time} can also be supplied as the total length of the simulation.}

\item{init}{Named vector of initial population sizes.
\itemize{
\item \code{N1}: Initial population size of species 1
\item \code{N2}: Initial population size of species 2
}}

\item{params}{Named vector of model parameters.
\itemize{
\item \code{r1}: Per capita growth rate of species 1
\item \code{r2}: Per capita growth rate of species 1
\item \code{a12}: Effect of species 2 on species 1
\item \code{a21}: Effect of species 1 on species 2
\item \code{b1}: Half-saturation constant for species 1
\item \code{b2}: Half-saturation constant for species 2
\item \code{d1}: Rate of self-limitation for species 1
\item \code{d2}: Rate of self-limitation for species 2
}}
}
\value{
A data frame of population sizes (\code{N1}, \code{N2}) simulated through time,
produced by solving \code{\link[=mutualism]{mutualism()}} with \code{\link[deSolve:ode]{deSolve::ode()}} given \code{time}, \code{init},
and \code{params}. This data frame has additional attributes recording the model
run (\code{model}), input parameter values (\code{params}) and population carrying
capacities calculated from the parameter values (\code{K}).
}
\description{
Given a vector of time, initial values, and parameters, this function runs
a phenomenological model of direct mutualism between two species as
described by Holland (2012). This model features a saturating functional
response for interspecific density dependence, which produces a stable
equilibrium where both mutualist species can coexist above their respective
population carrying capacities.
}
\note{
See the following for further context and details regarding this model:
Holland, N.J. (2012). Population dynamics of mutualism. Nature Education
Knowledge, 3, 2.
}
\examples{
# Define full time series and run model with default parameter values
run_mutualism(time = 0:10)

# Define length of simulation and run model with default parameter values
run_mutualism(time = 10)

# Run model with custom input values
tmax <- 10
start <- c(N1 = 100, N2 = 50)
pars <- c(r1 = 1, r2 = 1, a12 = 1.2, a21 = 0.8, b1 = 20, b2 = 20, d1 = 0.04, d2 = 0.02)
run_mutualism(time = tmax, init = start, params = pars)
}
\seealso{
\code{\link[=plot_mutualism_time]{plot_mutualism_time()}}, \code{\link[=plot_mutualism_portrait]{plot_mutualism_portrait()}}
}
