% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{run_ricker_model}
\alias{run_ricker_model}
\title{Ricker model of discrete population growth with a carrying capacity}
\usage{
run_ricker_model(N0 = 1, K = 100, r = 1.1, time = 100)
}
\arguments{
\item{N0}{initial population size of population}

\item{K}{Carrying capacity}

\item{r}{Population growth rate}

\item{time}{Number of time steps over which to project the model}
}
\value{
A data frame with colums time and Nt
}
\description{
Ricker model of discrete population growth with a carrying capacity
}
\examples{
run_ricker_model(N0 = 1, K = 100, r = 1.1, time = 100)
}
