% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smith_fretwell_functions.R
\name{run_smith_fretwell_model_sp2}
\alias{run_smith_fretwell_model_sp2}
\title{Add a new Smith-Fretwell curve to a ggplot object created with
run_smith_fretwell_model}
\usage{
run_smith_fretwell_model_sp2(
  existing_plot,
  Iymin = 6,
  Womax = 0.6,
  Iymax = 1000,
  shapeval = 0.02
)
}
\arguments{
\item{existing_plot}{ggplot object created with run_smith_fretwell_model that
has the first (base) curve}

\item{Iymin}{Minimum seed size for a viable offspring}

\item{Womax}{Maximum fitness that an offspring can achieve}

\item{Iymax}{Total amount of energy an adult has to invest into reproduction}

\item{shapeval}{slope of the Smith-Fretwell curve}
}
\description{
Add a new Smith-Fretwell curve to a ggplot object created with
run_smith_fretwell_model
}
\examples{
sf_species1 <- run_smith_fretwell_model(Iymin = 3, Womax = 0.6, Iymax = 1000)
run_smith_fretwell_model_sp2(sf_species1, Iymin = 6, Womax = 0.6, Iymax = 1000)
}
