% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_sink_dynamics.R
\name{run_source_sink}
\alias{run_source_sink}
\title{Simulate dynamics of the source and sink populations under Pulliam (1988)'s
model}
\usage{
run_source_sink(endtime, init, params)
}
\arguments{
\item{endtime}{a number to indicate at what timesetp to end the simulation}

\item{init}{a vector of initial population sizes of the source and sink
population}

\item{params}{a vector of model parameters}
}
\description{
Simulate dynamics of the source and sink populations under Pulliam (1988)'s
model
}
\examples{
# in this example, source and sink population start at the same size,
# the sink population dies out, but is later replenished by immigration
# from the source patch once all source breeding sites are occupied
run_source_sink(endtime = 50, init = c(n0Source = 100, n0Sink = 100),
params = c(pa = 0.6, pj = 0.15, betaSource = 3, betaSink = 1, NSource = 300))
}
\seealso{
\code{\link[=plot_source_sink]{plot_source_sink()}}  for plotting the size of the source and sink
population over time
}
