% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_population_growth.R
\name{run_structured_population_simulation}
\alias{run_structured_population_simulation}
\title{Simulate growth of a (st)age-structured population}
\usage{
run_structured_population_simulation(
  leslie_mat = matrix(c(0, 8, 1, 1, 0.4, 0, 0, 0, 0, 0.8, 0, 0, 0, 0, 0.1, 0), ncol =
    4, byrow = TRUE),
  init = c(10, 10, 10, 10),
  time = 100
)
}
\arguments{
\item{leslie_mat}{A leslie matrix}

\item{init}{A vector of initial abundances in each (st)age}

\item{time}{Length of time over which to run the simulation}
}
\value{
Returns a matrix with nrow = nrow(leslie_mat) and ncol = time
}
\description{
Simulate growth of a (st)age-structured population
}
\examples{
leslie_matrix <- matrix(c(0, 8,1, 1, 0.4,0,0,0,0,0.8,0,0,0,0,0.1,0),
ncol = 4, byrow = TRUE)
structured_pop_init <- c(10,10,10,10)
structured_pop_time <- 5
run_structured_population_simulation(leslie_mat = leslie_matrix, init =
structured_pop_init, time = structured_pop_time)
}
\seealso{
\code{\link[=plot_structured_population_size]{plot_structured_population_size()}},
\code{\link[=plot_structured_population_agedist]{plot_structured_population_agedist()}}, and
\code{\link[=plot_structured_population_lambda]{plot_structured_population_lambda()}} for plotting different aspects of the
population trajectory, and \code{\link[=plot_leslie_diagram]{plot_leslie_diagram()}} for plotting an
transition diagram based on the specified Leslie matrix
}
