% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinyapp_files.R
\name{shiny_abiotic_comp}
\alias{shiny_abiotic_comp}
\title{Launch the abiotic resource competition app}
\usage{
shiny_abiotic_comp()

shiny_abiotic_comp()
}
\description{
Launch the abiotic resource competition app

Launch the Smith-Fretwell app
}
