% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinyapp_files.R
\name{shiny_consumer_resource}
\alias{shiny_consumer_resource}
\title{Launch the consumer-resource dynamics app}
\usage{
shiny_consumer_resource()

shiny_consumer_resource()

shiny_consumer_resource()
}
\description{
Launch the consumer-resource dynamics app

Launch the biotic resource competition app

Launch the island biogeography app
}
