% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinyapp_files.R
\name{shiny_infectious_disease}
\alias{shiny_infectious_disease}
\title{Launch the infectious diseases models app}
\usage{
shiny_infectious_disease()
}
\description{
Launch the infectious diseases models app
}
