% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinyapp_files.R
\name{shiny_singlepop_discrete}
\alias{shiny_singlepop_discrete}
\title{Launch the single population dynamics in discrete time app}
\usage{
shiny_singlepop_discrete()
}
\description{
Launch the single population dynamics in discrete time app
}
