% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinyapp_files.R
\name{shiny_structured_pop}
\alias{shiny_structured_pop}
\title{Launch the structured population growth app}
\usage{
shiny_structured_pop()
}
\description{
Launch the structured population growth app
}
