% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_ODE_variants.R
\name{SIR_ft}
\alias{SIR_ft}
\title{SIR model with vital rates and frequency-dependent transmission}
\usage{
SIR_ft(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{vector of beta (infection rate), gamma (recovery rate),
m (natural birth/death rate), and v(vaccination rate)}
}
\description{
SIR model with vital rates and frequency-dependent transmission
}
\keyword{internal}
