% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectious_disease_dynamics.R
\name{SIR_no_vitals}
\alias{SIR_no_vitals}
\title{SIR model without vital rates}
\usage{
SIR_no_vitals(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{vector of beta (infection rate), gamma (recovery rate),
and v (vaccination rate)}
}
\description{
SIR model without vital rates
}
\keyword{internal}
