% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_ODE_variants.R
\name{SIS_ft}
\alias{SIS_ft}
\title{SIS model with vital rates and frequency-dependent transmission}
\usage{
SIS_ft(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{vector of beta (infection rate), m (natural birth/death rate),
and gamma (recovery rate)}
}
\description{
SIS model with vital rates and frequency-dependent transmission
}
\keyword{internal}
