% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_sink_dynamics.R
\name{source_sink}
\alias{source_sink}
\title{functions for Pullian's source sink dynamics model}
\usage{
source_sink(endtime, init, params)
}
\arguments{
\item{endtime}{how long to simulate the dynamics}

\item{init}{initial population size of population}

\item{params}{model parameters}
}
\description{
functions for Pullian's source sink dynamics model
}
