% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_resource_variants.R
\name{vector_field_input}
\alias{vector_field_input}
\title{generates a grid for plotting a vector field}
\usage{
vector_field_input(sim_df, pars_for_eq_func, vec_density = 20)
}
\arguments{
\item{sim_df}{data frame generated from run_XXX}

\item{pars_for_eq_func}{parameter values used to generate `sim_df`}

\item{vec_density}{density of grid to generate}
}
\value{
a data frame, with Hstart, Hend, Pstart, Pend,
and corresponding values of dH and dP for drawing vectors
}
\description{
generates a grid for plotting a vector field
}
