% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{bevertonholt_eqn}
\alias{bevertonholt_eqn}
\title{Equation that returns Nt+1 from Nt for the Beverton-Holt model}
\usage{
bevertonholt_eqn(Nt, Rd, K)
}
\arguments{
\item{K}{Carrying capacity}

\item{x}{Nt}

\item{r}{Population growth rate}
}
\description{
Equation that returns Nt+1 from Nt for the Beverton-Holt model
}
\keyword{internal}
