% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualism.R
\name{mutualism}
\alias{mutualism}
\title{Model of mutualism with saturating functional response}
\usage{
mutualism(time, init, params)
}
\arguments{
\item{time}{Vector of time units over which to run model, starting at 0.}

\item{init}{Named vector of initial population sizes.
\itemize{
\item \code{N1}: Initial population size of species 1
\item \code{N2}: Initial population size of species 2
}}

\item{params}{Named vector of model parameters.
\itemize{
\item \code{r1}: Per capita growth rate of species 1
\item \code{r2}: Per capita growth rate of species 1
\item \code{a12}: Effect of species 2 on species 1
\item \code{a21}: Effect of species 1 on species 2
\item \code{b1}: Half-saturation constant for species 1
\item \code{b2}: Half-saturation constant for species 2
\item \code{d1}: Rate of self-limitation for species 1
\item \code{d2}: Rate of self-limitation for species 2
}}
}
\value{
A list where the first and only element is a vector of derivatives
of the state variables (\code{N1}, \code{N2}) with respect to \code{time}, given \code{params}.
}
\description{
Used internally by \code{\link[=run_mutualism]{run_mutualism()}}.
}
\seealso{
\code{\link[=run_mutualism]{run_mutualism()}}
}
\keyword{internal}
