% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualism.R
\name{mutualism_vector_field}
\alias{mutualism_vector_field}
\title{Generate vector field for phase portrait of mutualism model}
\usage{
mutualism_vector_field(sim_df, vec.density = 20, vec.scale = 0.1)
}
\arguments{
\item{sim_df}{Data frame of mutualism model simulation returned by
\code{\link[=run_mutualism]{run_mutualism()}}.}

\item{vec.density}{Density of vectors, specified as a single numeric value
larger than 0. e.g. If \code{vec.density = 20} (default), a 20 \eqn{\times} 20
point grid evenly spaced along each axis is generated, where each grid
point is the starting point of a vector.}

\item{vec.scale}{Value to scale magnitude of vectors by, specified as a
single numeric value larger than 0. e.g. If \code{vec.scale = 0.1} (default) the
magnitude of each vector is 0.1 times the magnitude of the vector produced
by running \code{\link[=run_mutualism]{run_mutualism()}} for one time step given each starting point
(specified via \code{vec.density}) and \code{params} (inherited from \code{sim_df}).}
}
\value{
A data frame of vector start and end point values, collectively
representing the vector field for a phase portrait of the model of
mutualism with saturating functional response.
}
\description{
Used internally by \code{\link[=plot_mutualism_portrait]{plot_mutualism_portrait()}}.
}
\seealso{
\code{\link[=plot_mutualism_portrait]{plot_mutualism_portrait()}}
}
\keyword{internal}
