% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic_resource_competition.R
\name{plot_abiotic_comp_time}
\alias{plot_abiotic_comp_time}
\title{Plot population size over time for the aboitic resource competition model}
\usage{
plot_abiotic_comp_time(sim_df)
}
\arguments{
\item{sim_df}{matrix or data frame of simulations generated by
\link{run_abiotic_comp_model}}
}
\description{
Plot population size over time for the aboitic resource competition model
}
\examples{
params_vec <- c(S1 = 12, S2 = 12, r1 = 1.6, r2 = 1, k11 = 18, k12 = 4, k21 =
2, k22 = 14, m1 = .2, m2 = .2,c11 = .25, c12 = .08, c21 = .1, c22 = .2, a1 =
.5, a2 = .5)
sim_df_abioticrc <- run_abiotic_comp_model(time = seq(0,50), init = c(N1 =
10, N2 = 10, R1 = 20, R2 = 20), params = params_vec)
plot_abiotic_comp_time(sim_df_abioticrc)
}
\seealso{
\code{\link[=run_abiotic_comp_rstar]{run_abiotic_comp_rstar()}} for calculating both species' R* values
for both resource, \code{\link[=run_abiotic_comp_model]{run_abiotic_comp_model()}} for simulating the abiotic
resource competition model, and  \code{\link[=plot_abiotic_comp_portrait]{plot_abiotic_comp_portrait()}} for making
portrait plots of Resources 1 and 2 over time (including visualizations of
the ZNGIs)
}
