% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{plot_discrete_population_growth}
\alias{plot_discrete_population_growth}
\title{Plot the dynamics of a discrete population growth model}
\usage{
plot_discrete_population_growth(sim_df)
}
\arguments{
\item{sim_df}{data frame of discrete population growth}
}
\description{
Plot the dynamics of a discrete population growth model
}
\examples{
dl_params <- c(rd = 1.1, K = 1000)
sim_df_dl <- run_discrete_logistic_model(N0 = 1, params = dl_params, time = 100)
plot_discrete_population_growth(sim_df_dl)
bh_params <- c(rd = 1.1, K = 1000)
sim_df_bh <- run_beverton_holt_model(N0 = 1, params = bh_params, time = 100)
plot_discrete_population_growth(sim_df_bh)
ri_params <- c(rd = 1.1, K = 1000)
sim_df_ri <- run_ricker_model(N0 = 1, params = ri_params, time = 100)
plot_discrete_population_growth(sim_df_ri)
}
\seealso{
\code{\link[=run_discrete_exponential_model]{run_discrete_exponential_model()}}, \code{\link[=run_discrete_logistic_model]{run_discrete_logistic_model()}},
\code{\link[=run_beverton_holt_model]{run_beverton_holt_model()}}, \code{\link[=run_ricker_model]{run_ricker_model()}} for functions that
generate the \code{sim_df} objects plotted by this function
}
