% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualism.R
\name{plot_mutualism_time}
\alias{plot_mutualism_time}
\title{Plot population trajectories over time for mutualism model}
\usage{
plot_mutualism_time(sim_df)
}
\arguments{
\item{sim_df}{Data frame of mutualism model simulation returned by
\code{\link[=run_mutualism]{run_mutualism()}}.}
}
\value{
A ggplot object with trajectories of population sizes (\code{N1}, \code{N2}) plotted
against time. Population carrying capacities for each species, calculated
from model parameters, are also plotted.
}
\description{
Generate a plot of trajectories of population sizes over time for the model
of mutualism with saturating functional response.
}
\note{
The plot generated with this function is constrained to display only the
first quadrant (i.e. values of 0 and above) of the plane defined by \code{N} and
\code{time} because negative values of \code{N} or \code{time} do not make biological sense.
As a consequence, when \code{sim_df} is generated using negative values for \code{r1}
and/or \code{r2}, the corresponding lines for population carrying capacities will
not be visible in the plot because they will lie beyond the plot's limits.
}
\examples{
sim <- run_mutualism()
plot_mutualism_time(sim)
}
\seealso{
\code{\link[=run_mutualism]{run_mutualism()}}, \code{\link[=plot_mutualism_portrait]{plot_mutualism_portrait()}}
}
