% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predator_prey_dynamics.R
\name{plot_predprey_portrait}
\alias{plot_predprey_portrait}
\title{Function for plotting phase portrait of predator-prey model}
\usage{
plot_predprey_portrait(sim_df, params, vectors_field = FALSE, ...)
}
\arguments{
\item{sim_df}{data frame generated from run_XXX}

\item{params}{parameter values used to generate \code{sim_df}}

\item{vectors_field}{set to TRUE to see the vector field under the phase
portrait}

\item{...}{additional arguments to pass to \link{plot_vector_field}
(mostly this is the density of the vector field, set by vec_density)}
}
\description{
Function for plotting phase portrait of predator-prey model
}
\examples{
# Define parameters for the Rosenzweig-Macarthur model:
params <- c(r = .1, a = .01, e = .01, d = .001, K = 1000, T_h = .1)
sim_df <- run_predprey_model(200, init = c(H = 10, P = 5), params = params)
plot_predprey_portrait(sim_df = sim_df, params = params, vectors_field
= TRUE)
}
