% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predator_prey_dynamics.R
\name{plot_predprey_time}
\alias{plot_predprey_time}
\title{Function for plotting phase portrait of predator-prey model}
\usage{
plot_predprey_time(sim_df)
}
\arguments{
\item{sim_df}{data frame generated from run_XXX}
}
\description{
Function for plotting phase portrait of predator-prey model
}
\examples{
# Define parameters for the Rosenzweig-Macarthur model:
params <- c(r = .1, a = .01, e = .01, d = .01, K = 1000, T_h = .1)
sim_df <- run_predprey_model(200, init = c(H = 10, P = 5), params = params)
plot_predprey_time(sim_df = sim_df)
}
