% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_population_growth.R
\name{plot_structured_population_lambda}
\alias{plot_structured_population_lambda}
\title{Generate a trajectory of the total population size over time}
\usage{
plot_structured_population_lambda(pop_growth_matrix, leslie_mat = NULL)
}
\arguments{
\item{pop_growth_matrix}{population growth matrix, as generated by
`run_structured_population_simulation` (each row is an age class and each column
is a time step)}

\item{leslie_mat}{Leslie matrix used to generate the population trajectory (optional).
If the Leslie matrix is provided, the trajectory also includes lines indicating
the stable age distribution}
}
\value{
ggplot of total population size over time
}
\description{
Generate a trajectory of the total population size over time
}
