% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_population_growth.R
\name{plot_structured_population_size}
\alias{plot_structured_population_size}
\title{Generate a trajectory of population size in each (st)age over time}
\usage{
plot_structured_population_size(pop_growth_matrix)
}
\arguments{
\item{pop_growth_matrix}{population growth matrix, as generated by
`run_structured_population_simulation` (each row is an age class and each column
is a time step)}
}
\value{
ggplot object of population trajectory over time
}
\description{
Generate a trajectory of population size in each (st)age over time
}
