% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predator_prey_dynamics.R
\name{plot_vector_field}
\alias{plot_vector_field}
\title{generates a vector field}
\usage{
plot_vector_field(sim_df, params, vec_density = 20)
}
\arguments{
\item{sim_df}{data frame generated from run_predprey}

\item{params}{parameter values used to generate \code{sim_df}}

\item{vec_density}{density of grid to generate}
}
\description{
generates a vector field
}
\keyword{internal}
