% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_footer_apps.R
\name{print_app_footer}
\alias{print_app_footer}
\title{Print footer for Shiny apps}
\usage{
print_app_footer()
}
\description{
This function is only intended for use in EcoEvoApps' shiny apps.
}
\keyword{internal}
