% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic_resource_competition.R
\name{run_abiotic_comp_model}
\alias{run_abiotic_comp_model}
\title{Run the Tilman abiotic resource competition model}
\usage{
run_abiotic_comp_model(
  time = seq(0, 100, 0.1),
  init = c(N1 = 10, N2 = 10, R1 = 20, R2 = 20),
  params = c(S1 = 12, S2 = 12, r1 = 1.6, r2 = 1, k11 = 18, k12 = 4, k21 = 2, k22 = 14,
    m1 = 0.2, m2 = 0.2, c11 = 0.25, c12 = 0.08, c21 = 0.1, c22 = 0.2, a1 = 0.5, a2 = 0.5)
)
}
\arguments{
\item{time}{vector of time units over which to run model, starting from 0.
`time` can also be supplied as just the total length of the simulation (i.e. tmax)}

\item{init}{vector of initial population sizes for both species,
with names N1 and N2, AND initial resource pool sizes for both resources,
with names R1 and R2}

\item{params}{vector of model parameters (`S1`, `S2`, `r1`, `r2`, `k11`, `k12`,
`k21`, `k22`, `m1`, `m2`, `c11`, `c12`,
`c21`, `c22`, `a1`, `a1`)}
}
\description{
Run the Tilman abiotic resource competition model
}
\examples{
# Define full time series, and run model in terms of carrying capacities
# and relative competitive effects
run_abiotic_comp_model(time = seq(0,100,0.1),
 init = c(N1 = 10, N2 = 10, R1 = 20, R2 = 20),
 params = c(S1 = 12, S2 = 12, r1 = 1.6, r2 = 1,
k11 = 18, k12 = 4, k21 = 2, k22 = 14,
m1 = .2, m2 = .2,c11 = .25, c12 = .08,
c21 = .1, c22 = .2, a1 = .5, a2 = .5))
}
