% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{run_beverton_holt_model}
\alias{run_beverton_holt_model}
\title{Beverton Holt model}
\usage{
run_beverton_holt_model(N0 = 1, params = c(rd = 1.1, K = 100), time = 100)
}
\arguments{
\item{N0}{initial population size of population}

\item{params}{a vector of carrying capacity (K) and discrete growth rate (rd)}

\item{time}{Number of time steps over which to project the model}
}
\value{
A data frame with columns time and Nt
}
\description{
Beverton Holt model
}
\examples{
run_beverton_holt_model(N0 = 1, params = c(rd = 1.1, K = 100), time = 100)
}
