% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{run_discrete_exponential}
\alias{run_discrete_exponential}
\title{Discrete exponential growth model}
\usage{
run_discrete_exponential(N0 = 1, lambda = 1.1, time = 100)
}
\arguments{
\item{N0}{initial population size of population}

\item{lambda}{discrete population growth rate}

\item{time}{Number of time steps over which to project the model}
}
\value{
A data frame with colums time and Nt
}
\description{
Discrete exponential growth model
}
\examples{
run_discrete_exponential(N0 = 1, lambda = 1.1, time = 100)
}
