% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_continuous.R
\name{run_exponential_model}
\alias{run_exponential_model}
\title{Run exponential growth model}
\usage{
run_exponential_model(time = 10, init = c(N1 = 1), params = c(r = 0.1))
}
\arguments{
\item{time}{vector of time units over which to run model, starting from 0.
\code{time} can also be supplied as just the total length of the simulation
(i.e. tmax)}

\item{init}{initial population size of population, in a vector with name \code{N1}}

\item{params}{intrinsic growth rate r, in a vector with name \code{r}}
}
\description{
Run exponential growth model
}
\examples{
run_exponential_model(time = 0:10, init = c(N1 = 1), params = c(r = .1))
run_exponential_model(time = 10, init = c(N1 = 1), params = c(r = .1))
}
\seealso{
\code{\link[=run_logistic_model]{run_logistic_model()}} for simulating the dynamics of a population
with logistic growth to a carrying capacity, and
\code{\link[=run_discrete_exponential_model]{run_discrete_exponential_model()}} for a simulating exponential growth in
discrete time
}
