% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_continuous.R
\name{run_exponential_model}
\alias{run_exponential_model}
\title{Run exponential growth model}
\usage{
run_exponential_model(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model, starting from 0.
`time` can also be supplied as just the total length of the simulation (i.e. tmax)}

\item{init}{initial population size of population, in a vector with name `N1`}

\item{params}{intrinsic growth rate r, in a vector with name `r`}
}
\description{
Run exponential growth model
}
\examples{
run_exponential_model(time = 0:100, init = c(N1 = 1), params = c(r = .1))
run_exponential_model(time = 100, init = c(N1 = 1), params = c(r = .1))
}
