% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_ODE_variants.R
\name{run_infectiousdisease_model}
\alias{run_infectiousdisease_model}
\title{Run infectious disease models models}
\usage{
run_infectiousdisease_model(time, init, params, model_type)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{vector of beta (infection rate), m (natural birth/death rate),
and gamma (recovery rate)}

\item{model_type}{which type of model to run? (should be one of
\code{SIR}, \code{SIR_ft}, \code{SEIR}, \code{SEIR_ft}, \code{SIRD}, \code{SIRD_ft}, \code{SIS}, or \code{SIS_ft})}
}
\description{
Run infectious disease models models
}
\examples{
# Run the SIR model
params_vec <- c(m = .1, beta = .01, v = .2, gamma = 0)
init_vec <- c(S = 100, I = 1, R = 0)
time_vec <- seq(0, 10, 0.1)
run_infectiousdisease_model(time = time_vec, init = init_vec, params = params_vec, model_type = "SIR")
}
