% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_ODE_variants.R
\name{run_infectiousdisease_model}
\alias{run_infectiousdisease_model}
\title{Run infectious disease models models}
\usage{
run_infectiousdisease_model(time, init, params, model_type)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{vector of beta (infection rate), m (natural birth/death rate),
and gamma (recovery rate)}

\item{model_type}{which type of model to run? (should be one of
`SIR`, `SIR_ft`, `SEIR`, `SEIR_ft`, `SIRD`, `SIRD_ft`, `SIS`, or `SIS_ft`)}
}
\description{
Run infectious disease models models
}
