% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer_resource_variants.R
\name{run_predprey_model}
\alias{run_predprey_model}
\title{Run predator-prey model}
\usage{
run_predprey_model(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{pars}{intrinsic growth rate r, a, e, d, T_h, K}
}
\description{
Run predator-prey model
}
