% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_population_discrete.R
\name{run_ricker_model}
\alias{run_ricker_model}
\title{Ricker model of discrete population growth with a carrying capacity}
\usage{
run_ricker_model(N0 = 1, params = c(rd = 1.1, K = 100), time = 100)
}
\arguments{
\item{N0}{initial population size of population}

\item{params}{a vector of carrying capacity (K) and discrete growth rate (rd)}

\item{time}{Number of time steps over which to project the model}
}
\value{
A data frame with columns time and Nt
}
\description{
Ricker model of discrete population growth with a carrying capacity
}
\examples{
run_ricker_model(N0 = 1, params = c(K = 100, rd = 1.1), time = 10)
}
