% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinyapp_files.R
\name{shiny_XXX}
\alias{shiny_XXX}
\alias{shiny_singlepop_continuous}
\alias{shiny_singlepop_discrete}
\alias{shiny_lvcompetition}
\alias{shiny_consumer_resource}
\alias{shiny_biotic_comp}
\alias{shiny_infectious_disease}
\alias{shiny_metapopulation}
\alias{shiny_structured_pop}
\alias{shiny_island_biogeo}
\alias{shiny_abiotic_comp}
\alias{shiny_smith_fretwell}
\title{Run the shiny apps for various models}
\usage{
shiny_singlepop_continuous()

shiny_singlepop_discrete()

shiny_lvcompetition()

shiny_consumer_resource()

shiny_biotic_comp()

shiny_infectious_disease()

shiny_metapopulation()

shiny_structured_pop()

shiny_island_biogeo()

shiny_abiotic_comp()

shiny_smith_fretwell()
}
\arguments{
\item{language}{Two-letter vector indicating the language}
}
\description{
Run the shiny apps for various models
}
