% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinyapp_files.R
\name{shiny_XXX}
\alias{shiny_XXX}
\alias{shiny_singlepop_continuous}
\alias{shiny_population_growth_discrete}
\alias{shiny_lvcomp}
\alias{shiny_predprey}
\alias{shiny_biotic_comp}
\alias{shiny_infectious_disease}
\alias{shiny_source_sink}
\alias{shiny_structured_population}
\alias{shiny_ibiogeo_model}
\alias{shiny_abiotic_comp}
\alias{shiny_smith_fretwell}
\alias{shiny_mutualism}
\title{Run the shiny apps for various models}
\usage{
shiny_singlepop_continuous(language = "en")

shiny_population_growth_discrete(language = "en")

shiny_lvcomp(language = "en")

shiny_predprey(language = "en")

shiny_biotic_comp(language = "en")

shiny_infectious_disease(language = "en")

shiny_source_sink()

shiny_structured_population(language = "en")

shiny_ibiogeo_model(language = "en")

shiny_abiotic_comp(language = "en")

shiny_smith_fretwell()

shiny_mutualism(language = "en")
}
\arguments{
\item{language}{Two-letter vector indicating the language
For most models, apps are available in the following languages:
english (\code{language = "en"}), spanish (\code{language = "es"}),
chinese (\code{language = "ch"}), turkish (\code{language = "tr"}),
and portuguese (\code{language = "pt"}).}
}
\description{
Run the shiny apps for various models
}
