% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic_resource_competition.R
\name{tilman_comp_essential}
\alias{tilman_comp_essential}
\title{Competition between two species for two essential resources}
\usage{
tilman_comp_essential(time, init, params)
}
\arguments{
\item{time}{vector of time units over which to run model}

\item{init}{initial population size of population}

\item{params}{vector of parameters: intrinsic growth rate r for the resource, d1 and d2 for intrinsic mortality rate of sp1 and sp2; a1 and a2 for resource uptake rate of sp1 and sp2}
}
\description{
Competition between two species for two essential resources
}
