% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predator_prey_dynamics.R
\name{vector_field_input}
\alias{vector_field_input}
\title{Generate a grid for plotting a vector field under predator-prey trajectories}
\usage{
vector_field_input(sim_df, params, vec_density = 20)
}
\arguments{
\item{sim_df}{data frame generated from run_XXX}

\item{params}{parameter values used to generate \code{sim_df}}

\item{vec_density}{density of grid to generate}
}
\value{
a data frame, with Hstart, Hend, Pstart, Pend,
and corresponding values of dH and dP for drawing vectors
}
\description{
Generate a grid for plotting a vector field under predator-prey trajectories
}
\keyword{internal}
